backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# OCR libraries
easyocr==1.7.1
pdf2image==1.16.3
Pillow==10.1.0

# HTTP client for Cerebras API
httpx==0.25.1

# Image processing dependencies (for pdf2image)
# Note: pdf2image also requires poppler to be installed on the system
# Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
# Linux: sudo apt-get install poppler-utils
# macOS: brew install poppler
